Rename nfdi mod to aruna;
