+++ services/src/error.rs
@@ -1,10 +1,10 @@
 use crate::api::model::datatypes::{
     DataProviderId, DatasetId, LayerId, SpatialReference, SpatialReferenceOption, TimeInstance,
 };
+#[cfg(feature = "nfdi")]
+use crate::datasets::external::aruna::error::ArunaProviderError;
 #[cfg(feature = "ebv")]
 use crate::datasets::external::netcdfcf::NetCdfCf4DProviderError;
-#[cfg(feature = "nfdi")]
-use crate::datasets::external::nfdi::error::ArunaProviderError;
 use crate::handlers::ErrorResponse;
 use crate::{layers::listing::LayerCollectionId, workflows::workflow::WorkflowId};
 use actix_web::http::StatusCode;
